requirements.txt
Flask==1.1.2
gunicorn==20.0.4
pylint==2.6.0
pylint-quotes==0.2.1
PyYAML==5.3.1
requests==2.24.0
